Move subresource tests file under compute pkg
