+++ tests/BUILD.bazel
@@ -69,15 +69,12 @@ go_test(
         "hyperv_test.go",
         "kubectl_test.go",
         "mdev_configuration_allocation_test.go",
-        "pausing_test.go",
         "pdb_test.go",
         "pool_test.go",
         "portforward_test.go",
         "replicaset_test.go",
         "security_features_test.go",
-        "softreboot_test.go",
         "stability_test.go",
-        "subresource_api_test.go",
         "swap_test.go",
         "sysprep_test.go",
         "tests_suite_test.go",
@@ -121,7 +118,6 @@ go_test(
         "//pkg/hooks/v1alpha1:go_default_library",
         "//pkg/hooks/v1alpha2:go_default_library",
         "//pkg/hooks/v1alpha3:go_default_library",
-        "//pkg/instancetype:go_default_library",
         "//pkg/libdv:go_default_library",
         "//pkg/libvmi:go_default_library",
         "//pkg/libvmi/cloudinit:go_default_library",
@@ -157,6 +153,7 @@ go_test(
         "//staging/src/kubevirt.io/client-go/subresources:go_default_library",
         "//tests/clientcmd:go_default_library",
         "//tests/compute:go_default_library",
+        "//tests/compute/subresources:go_default_library",
         "//tests/console:go_default_library",
         "//tests/containerdisk:go_default_library",
         "//tests/decorators:go_default_library",
@@ -174,7 +171,6 @@ go_test(
         "//tests/libconfigmap:go_default_library",
         "//tests/libinfra:go_default_library",
         "//tests/libinstancetype:go_default_library",
-        "//tests/libinstancetype/builder:go_default_library",
         "//tests/libkubevirt:go_default_library",
         "//tests/libkubevirt/config:go_default_library",
         "//tests/libmigration:go_default_library",
@@ -230,10 +226,8 @@ go_test(
         "//vendor/k8s.io/apimachinery/pkg/fields:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/intstr:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/json:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/rand:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/validation/field:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/wait:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
         "//vendor/k8s.io/client-go/informers:go_default_library",
