+++ tests/compute/subresources/BUILD.bazel
@@ -1,0 +1,43 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "pausing.go",
+        "softreboot.go",
+        "subresource_api.go",
+    ],
+    importpath = "kubevirt.io/kubevirt/tests/compute/subresources",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/instancetype:go_default_library",
+        "//pkg/libvmi:go_default_library",
+        "//pkg/pointer:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/api/instancetype:go_default_library",
+        "//staging/src/kubevirt.io/api/instancetype/v1beta1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
+        "//staging/src/kubevirt.io/client-go/kubevirt/typed/core/v1:go_default_library",
+        "//tests/compute:go_default_library",
+        "//tests/console:go_default_library",
+        "//tests/decorators:go_default_library",
+        "//tests/framework/kubevirt:go_default_library",
+        "//tests/framework/matcher:go_default_library",
+        "//tests/libinstancetype/builder:go_default_library",
+        "//tests/libpod:go_default_library",
+        "//tests/libvmifact:go_default_library",
+        "//tests/libvmops:go_default_library",
+        "//tests/libwait:go_default_library",
+        "//tests/testsuite:go_default_library",
+        "//vendor/github.com/google/goexpect:go_default_library",
+        "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
+        "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/intstr:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/validation/field:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
+        "//vendor/k8s.io/client-go/rest:go_default_library",
+    ],
+)
