+++ tests/tests_suite_test.go
@@ -43,6 +43,7 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 
 	_ "kubevirt.io/kubevirt/tests/compute"
+	_ "kubevirt.io/kubevirt/tests/compute/subresources"
 	_ "kubevirt.io/kubevirt/tests/guestlog"
 	_ "kubevirt.io/kubevirt/tests/hotplug"
 	_ "kubevirt.io/kubevirt/tests/infrastructure"
